move brain tools to separated files
